+++ internal/usecase/calendar_impl.go
@@ -1,0 +1,29 @@
+package usecase
+
+import (
+	"bonded/internal/models"
+	"context"
+	"github.com/google/uuid"
+)
+
+func (u *usecase) Calendar() CalendarUsecase {
+	return u.calendarUsecase
+}
+
+func (u *calendarUsecase) CreateCalendar(ctx context.Context, calendar *models.Calendar) error {
+	calendar.ID = uuid.New().String()
+	return u.calendarRepo.Create(ctx, calendar)
+}
+
+func (u *calendarUsecase) EditCalendar(ctx context.Context, calendar *models.Calendar) error {
+	return u.calendarRepo.Edit(ctx, calendar)
+}
+
+func (u *calendarUsecase) DeleteCalendar(ctx context.Context, calendarID string) error {
+	return u.calendarRepo.Delete(ctx, calendarID)
+}
+
+func (u *calendarUsecase) FindCalendars(ctx context.Context, userID string) ([]*models.Calendar, error) {
+	return u.calendarRepo.FindByUserID(ctx, userID)
+}
+
